requirements.txt
Flask
psycopg2-binary 
psycopg2
python-dotenv